Moved types folder to root
