+++ declarations.d.ts
@@ -1,7 +1,7 @@
 interface Window {
   sql: {
     planets: () => Promise<
-      import('./src/app/core/types/PlanetAffiliation').PlanetAffiliationJSON[]
+      import('./src/app/types/PlanetAffiliation').PlanetAffiliationJSON[]
     >;
   };
 }
