+++ src/app/types/AffiliationJson.ts
